plugins/adl_ftp_plugin/requirements/base.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=base.txt base.in
#
anyascii==0.3.2
    # via wagtail
asgiref==3.8.1
    # via django
beautifulsoup4==4.12.3
    # via wagtail
certifi==2024.8.30
    # via requests
charset-normalizer==3.4.0
    # via requests
defusedxml==0.7.1
    # via willow
django==5.1.3
    # via
    #   django-filter
    #   django-modelcluster
    #   django-permissionedforms
    #   django-polymorphic
    #   django-taggit
    #   django-treebeard
    #   djangorestframework
    #   laces
    #   wagtail
django-filter==24.3
    # via wagtail
django-modelcluster==6.3
    # via wagtail
django-permissionedforms==0.1
    # via wagtail
django-polymorphic==3.1.0
    # via -r base.in
django-taggit==5.0.1
    # via wagtail
django-treebeard==4.7.1
    # via wagtail
djangorestframework==3.15.2
    # via wagtail
draftjs-exporter==5.0.0
    # via wagtail
et-xmlfile==2.0.0
    # via openpyxl
filetype==1.2.0
    # via willow
idna==3.10
    # via requests
l18n==2021.3
    # via wagtail
laces==0.1.1
    # via wagtail
openpyxl==3.1.5
    # via wagtail
pillow==10.4.0
    # via
    #   pillow-heif
    #   wagtail
pillow-heif==0.20.0
    # via willow
python-dateutil==2.9.0.post0
    # via -r base.in
pytz==2024.2
    # via
    #   django-modelcluster
    #   l18n
requests==2.32.3
    # via wagtail
six==1.16.0
    # via
    #   l18n
    #   python-dateutil
soupsieve==2.6
    # via beautifulsoup4
sqlparse==0.5.2
    # via django
telepath==0.3.1
    # via wagtail
urllib3==2.2.3
    # via requests
wagtail==6.1.2
    # via -r base.in
willow[heif]==1.9.0
    # via
    #   wagtail
    #   willow
